fix: Refactor AWS modules to use modules/aws directory
